requirements.txt
torch==2.0.1
torchvision==0.15.2
numpy==1.23.5
scikit-learn==1.2.2
pandas==1.5.3
Pillow==9.5.0
pydantic
fastapi
